Finish first verstion of styleguide
We have three sections: colors, elements and typography.

Colors: all the site colors.

Elements: a little general, but all the standard elements like car
tiles, image elements, buttons, etc. Now it is incomplete. At the time
we add more elements to that page, we are probably going to split it
into two or more pages.

Typohtaphy: it is empty by now but it will have all our fonts and main
elements sizes.